Record counters with three different levels of granularity
The most recent data we want to keep with a per-minute resultion, but we
only want to keep that data around for a single day.

Older data we only need to keep with a resolution of a day; that is,
when looking at data from 12 months ago we aren't going to be analyzing
at the level of minutes and hours but rather days.

Rather than keeping a single counter when recording an event, and then
having to worry about migrating it to the longer-term, lower-resolution
data store, we just store all three levels of granularity at once, but
using different identifying keys and different expiries.