Initial support for simple querying
We can query for event counts by visiting:

   /count/my-fancy-event

   (returns minute-resolution data for the last hour)

or:

  /count/my-fancy-event/hours/5

  (returns hour-resolution data over the last five hours)

For unknown events or events which haven't been tracked yet we'll just
return 0. We'll return an error response if we get an unknown
"granularity" param (accepted values are "minutes", "hours" and "days").
Likewise we'll return an error response if we get a quantity param that
exceeds the expiry threshold for the given granularity; this is to stop
rogue requests like "show me minute-resolution data for the last 9,000
trillion minutes" from bringing down the server.