Teach /track method to accept optional quantity param
There are some places in our application where we process jobs in
batches. In some cases we'll want to sent individual track notifications
for those events, but there are other cases where this makes less sense:
for example, when people send out Cause invitations on our platform,
it's actually Facebook that sends the invites and then tells us how many
were sent.

For cases like this, it seems much more sensible to just tell Duckweed
"45 of this thing just happened", rather than send 45 separate
notifications in a loop.

This will also open up the possibility of using Duckweed to answer not
only questions like "what happened?", but also "how much of it
happened?". In other words, we can use it to monitor not just how many
donations were made, but how many dollars were donated; without the
ability to increment in batches, we'd be forced to do one notification
for every dollar donated.