Don't examine the current bucket when reporting.
Ignore the current partially-filled bucket when getting data for
geckoboard et al, so as to avoid the scary-looking precipitous
falloff at the end of every graph.

This commit examines the correctly-sized window of buckets, shifted
one (minute|hour|day) back in time.

Changes to existing tests and new test to account for that change.