Support an 'offset' param to pick range of data;  use that for 'skip latest'.
Duckweed now honors an 'offset' optional param that picks how many buckets
ago to return the range for.  This could be used for "show me by the minute,
for six hours ago", for instance

Offset defaults to 1, which is the mechanism by which we skip the most
recent bucket to avoid steep cliffs on the right side of graphs.