Speed up tests by ~90%.
Make the helper method #event_happened much more efficient when the
event count is > 1. Prior to this change, event_happened(:times => N)
would make N POST requests to the app; now, it makes just one and sets
the :quantity parameter to N.

This slowness occurred because #event_happened predates the :quantity
parameter, so repeated requests was the only way to go.