Add permission levels to tokens ('r' and 'w').
POST /track requires a 'w' token because it actually updates the
database.

GET /count, GET /histogram, etc. require an 'r' token because they
just read data from the database.

This way, you can give an r-token to Geckoboard to get some pretty
graphs, but they can't use that to insert data (not that they would;
I'm sure the folks at Geckoboard are fine upstanding citizens).

Duckweed::Token doesn't assign any meanings to the permissions stored
in the token; it's just a string. Duckweed::App is the thing that
assigns semantics to 'r' and 'w'. Should we want a finer-grained
permission system in the future, Duckweed::Token is ready for
it. (Perhaps 'c' for GET /count, 'h' for GET /histogram, et cetera;
seems like overkill, but you never know.)