Bump up Redis timeout to 30 seconds.
It seems that every hour or so, probably when Redis is rewriting its
append-only file in the background, certain operations take longer
than the default timeout of 5 seconds, so we get a bunch of failures
and wind up both losing data and spamming Airbrake.

It shouldn't hurt anything to have a longer timeout. If it's a POST to
/track, you should be doing that in the background anyway. If you're
requesting data, then it's better for us to be slow than to return an
error.