Add README
In order to open-source Duckweed, we need a basic README. This one
should cover most things of interest. I did not yet document the new
"grouped event" feature that Lann added the other day, as I think we
should "dogfood" it for a while (ie. try it out and iron out any kinks)
before we commit to making it a permanent part of the API.

Related story:
  https://www.pivotaltracker.com/story/show/13888561

Change-Id: Ie8c428ae75278a08903cfbaa038658358052ca7d
Reviewed-on: https://gerrit.causes.com/1107
Tested-by: Greg Hurrell <greg@causes.com>
Reviewed-by: Kristján Pétursson <kristjan@causes.com>